requirements.txt
PyYAML
fastapi
uvicorn[standard]
pydantic
langchain-core
langchain-google-vertexai
google-cloud-aiplatform
chromadb
python-dotenv